snmp: shuffle code around into separate files per type
